Fix broken Windows build
