+++ v3/dump_unix.go
@@ -5,7 +5,6 @@ package luddite
 import (
 	"os"
 	"os/signal"
-	"runtime"
 	"sync"
 	"syscall"
 )
@@ -24,15 +23,4 @@ func dumpGoroutineStacks() {
 		}()
 		signal.Notify(sigs, syscall.SIGUSR1)
 	})
-}
-
-func goroutineStack(all bool) []byte {
-	buf := make([]byte, 1024)
-	for {
-		n := runtime.Stack(buf, all)
-		if n < len(buf) {
-			return buf[:n]
-		}
-		buf = make([]byte, 2*len(buf))
-	}
 }
